*All logic done (except Save/Load)

+++ Codeme_adventureGame_BartoszK/Assets/Scripts/EnemyPrefs.cs
@@ -9,6 +9,6 @@ public class EnemyPrefs : ScriptableObject
     public float speed;
     public int changeDirectionRate;
     public float energy;
-    public float bitePower;
+    public int bitePower;
 
 }
